+++ src/test/java/com/monkey/properties/file/testing/PropertiesFileWriterTest.java
@@ -11,7 +11,6 @@ import java.util.Properties;
 import org.testng.annotations.Test;
 
 import com.tvd12.properties.file.exception.PropertiesFileException;
-import com.tvd12.properties.file.writer.Base64FileWriter;
 import com.tvd12.properties.file.writer.BaseFileWriter;
 
 public class PropertiesFileWriterTest {
@@ -23,28 +22,11 @@ public class PropertiesFileWriterTest {
         new BaseFileWriter().write(properties, "hello.output_test");
     }
     
-    @Test
-    public void testBase64() throws PropertiesFileException {
-        Properties properties = new Properties();
-        properties.setProperty("hello", "word");
-        new Base64FileWriter().write(properties, "hello_base64.test_output");
-    }
-    
     @Test(expectedExceptions = {PropertiesFileException.class})
     public void writeToOuputStreamInvalidCase() throws PropertiesFileException, IOException {
         Properties properties = mock(Properties.class);
         doThrow(IOException.class).when(properties).store(any(OutputStream.class), any(String.class));
         new BaseFileWriter().write(properties);
     }
-    
-//    @Test(expectedExceptions = {PropertiesFileException.class})
-//    public void writeToFileInvalidCase() throws PropertiesFileException, IOException {
-//        Properties properties = new Properties();
-//        File file = mock(File.class);
-//        when(file.exists()).thenReturn(true);
-//        when(file.isDirectory()).thenReturn(true);
-//        when(file.getPath()).thenReturn("hello");
-//        new BaseFileWriter().write(properties, file);
-//    }
     
 }